+++ src/cqterrain/crystal/__init__.py
@@ -1,1 +1,2 @@
 from .crystal_random import crystal_random
+from .CrystalWall import CrystalWall