+++ example/minibase/circle_uneven.py
@@ -1,0 +1,19 @@
+import cadquery as cq
+from cqterrain.minibase import circle_uneven
+
+ex_base = circle_uneven(
+    diameter = 45,
+    base_height = 3,
+    taper = -1,
+    render_magnet = True,  
+    magnet_diameter = 3, 
+    magnet_height = 2,
+    detail_height = 3,
+    uneven_height = 4,
+    peak_count = (9,10),
+    segments = 6,
+    seed="red"
+)
+
+#show_object(ex_base)
+cq.exporters.export(ex_base,'stl/minibase_circle_uneven.stl')