+++ src/cqterrain/minibase/__init__.py
@@ -1,5 +1,6 @@
 from .make_magnet_outline import make_magnet_outline
 from .circle import circle
+from .circle_uneven import circle_uneven
 from .ellipse import ellipse
 from .rectangle import rectangle
 from .slot import slot