+++ src/cqterrain/book/__init__.py
@@ -1,0 +1,3 @@
+from .book import book
+from .book_random import book_random
+from .books import books