+++ src/cqterrain/crystal/CrystalWall.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import cadquery as cq
 from cadqueryhelper import Base
 from cqterrain.minibase import slot_uneven