+++ changes.md
@@ -1,4 +1,10 @@
 ## Main wip
+## 2.8.0
+* Added book package
+  * Added book
+  * Added book_random
+  * Added books
+  * Updated README.md
 
 ## 2.7.0
 * Fixed source file paths in damage.md